requirements.txt
numpy==1.17.2
numpydoc==0.9.1
opencv-python==4.2.0.34
pathlib2==2.3.5
Pillow==6.2.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
torch==1.4.0
torchvision==0.5.0
